Fix for enumeration error while checking beacons
Fix for "collection was modified; enumeration operation may not execute" exception in CheckBeacons() that is caught and causes no problems, but needlessly writes to the console when it happens. This is due to beacons being added to or removed from NaniteConstructionManager.BeaconList while the script is running in parallel. Adding ToList will create a copy and prevent it from being modified. Adding a few null checks to compensate.