SP compatibility, MP sync fixes
Fix multiplayer syncing for control panel info box, emissives and state. Force m_factorystate to be sent to client when the state checks, but changed it to only send to clients in sync range. Processing for finding clients in sync range has been moved to parallel. Allowed compatibility for emissives in singleplayer, since no client-server messages are exchanged.