Call .Append less often, use Parallel.Start
We're going to use Parallel.Start rather than Parallel.StartBackground in order to save on the amount of spawned threads in the game, which might cause a traffic jam. Parallel.Start uses a pool of always-running internal threads and is better for smaller operations. We'll still use Parallel.StartBackground for huge operations like building inventory or scanning blocks/targets. Reduced the calls to details.Append when strings could easily just be concatenated. Put factory status logging into parallel to save on performance.