Smooth out animation transition for new center orb
Animations were janky because ProcessState is only called once every two seconds, so the orb would just freeze at the top of the spool for a second or two. Now, once the top of the spool is reached, ProcessState is forced to run so the spinning "active" animation begins immediately.