Get mining up and running
Gets the material, cuts it out of the voxel and adds it to the inventory. NOT performance friendly. This is a preliminary build to get it functional and open it up to the rest of you for contributions. Will do more work soon. Currently up and running on the test server.