Brand new block scanning status, performance
Scanned blocks are now cached and held while they are scanned at a rate of 100 blocks per second to keep server performance smoother and less spiky. This value will be configurable. Various performance tweaks, and a brand new console status that displays the total blocks scanned and the processing status with a percentage.