Don't clear potential targets until later
Was clearing potential targets before needed inventory was found. This fixes that, and adds support for a new, conglomerate "Count" variable that is separate from the actual current potential targets (which changes every 5 seconds if a scan includes more than 500 blocks). This makes more accurate reporting to console and better factory state handling.