Assembler will now ignore existing queue
Assembler would previously reset player queues (insetQueueItem does not add on top of existing queues. It might have in the past. It resets any existing queues and replaces them). Now, if that same item is already queued by a player or from a previous ProcessAssemblerQueue pass, we will let that queue completely finish before adding more, as it was just needlessly removing and replacing queues over and over and slowing down the work.